move persistence imple to new Fragment
